refactor: CardNameInput 파일 jsx -> tsx로 변경
